+++ resolwe/elastic/management/commands/__init__.py
@@ -8,4 +8,8 @@ Elastic app includes following Django management commands:
 
 .. automodule:: resolwe.elastic.management.commands.elastic_index
 
+.. automodule:: resolwe.elastic.management.commands.elastic_mapping
+
+.. automodule:: resolwe.elastic.management.commands.elastic_purge
+
 """